change id to long from int to match entity

+++ tsmoreland.objecttracker/App/src/main/java/tsmoreland/objecttracker/App/models/ObjectSummaryDataTransferObject.java
@@ -14,20 +14,20 @@ package tsmoreland.objecttracker.app.models;
 
 public class ObjectSummaryDataTransferObject {
 
-    private int id;
+    private long id;
     private String name;
 
-    public ObjectSummaryDataTransferObject(int id, String name) {
+    public ObjectSummaryDataTransferObject(long id, String name) {
         this.id = id;
         this.name = name;
     }
     public ObjectSummaryDataTransferObject() {
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
     public String getName() {
